% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_list}
\alias{add_list}
\title{Add one or more packages listed in a text file to a CRAN-like repository}
\usage{
add_list(list_file, ...)
}
\arguments{
\item{list_file}{A file path containing a list of R package references, one
per line.}

\item{...}{Additional arguments passed to \link{add_pkg}.}
}
\description{
Add one or more packages listed in a text file to a CRAN-like repository
}
