% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_repo}
\alias{add_repo}
\title{Add all packages from a CRAN-like repository to a package repository}
\usage{
add_repo(repos = ppm_config$cran_mirror, skip = FALSE, ...)
}
\arguments{
\item{repos}{A character vector containing the base URL(s) of CRAN-like R
package repositories. Defaults to the Posit Package Manager CRAN mirror.}

\item{skip}{A character string containing a regular expression matching names
of packages to skip. Defaults to \code{FALSE}, meaning keep all packages.}

\item{...}{
  Arguments passed on to \code{\link[=add_pkg]{add_pkg}}
  \describe{
    \item{\code{repo_dir}}{The package repository directory. Defaults to \code{"./repo"}.}
    \item{\code{remotes}}{A character vector of package references to prefer as a remote
source. Defaults to \code{NA}, meaning prefer a built-in list of references to
packages pre-modified for use with webR.}
    \item{\code{dependencies}}{Dependency specification for packages to additionally
add to the repository. Defaults to \code{FALSE}, meaning no additional packages.
Use \code{NA} to install only hard dependencies whereas \code{TRUE} installs all
optional dependencies as well. See \link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies}
for details.}
  }}
}
\description{
Downloads and builds all available source R packages from the R package
repositories given by \code{repos}, compiling each package for use with
WebAssembly and webR. The resulting WebAssembly binary packages are added to
the repository directory \code{repo_dir}. The repository directory will be created
if it does not already exist.
}
