% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build one or more R packages for WebAssembly}
\usage{
build(packages, remotes = NULL, out_dir = ".", dependencies = FALSE)
}
\arguments{
\item{packages}{A character vector of one or more package references.}

\item{remotes}{A character vector of package references to prefer as a remote
source. If \code{NULL}, use a built-in list of references to packages
pre-modified for use with webR.}

\item{out_dir}{The output directory. Defaults to \code{"."}.}

\item{dependencies}{Dependency specification for packages to additionally
build. Defaults to \code{FALSE}, meaning no additional packages. See
\link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies} for details.}
}
\description{
Downloads and builds the \link[pkgdepends:pkg_refs]{R package references} given
by \code{packages}, compiling each package for use with WebAssembly and webR. The
resulting WebAssembly binary packages are written to \code{out_dir}.
}
