% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{file_packager}
\alias{file_packager}
\title{Build an Emscripten VFS image}
\usage{
file_packager(in_dir, out_dir = "./vfs", out_name = NULL)
}
\arguments{
\item{in_dir}{A directory path to be packaged into the Emscripten VFS image.}

\item{out_dir}{A directory path to where the output files will be written.}

\item{out_name}{The output Emscripten VFS image file name base. If \code{NULL},
defaults to the final component of the input directory path.}
}
\description{
Uses Emscripten's \code{file_packager} tool to build an Emscripten VFS image that
can be mounted by webR. The image may contain arbitrary data that will be
made available for use by the WebAssembly R process once mounted.
}
\details{
Outputs two files (named by \code{out_name}) in the \code{out_dir} directory: a data
file with extension \code{".data"}, and a metadata file with extension
\code{".js.metadata"}. Both files should be hosted on the web so that their URL
can be passed to Emscripten for mounting.
}
