% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_library}
\alias{make_library}
\title{Create an R package library}
\usage{
make_library(repo_dir = "./repo", lib_dir = "./lib", strip = NULL)
}
\arguments{
\item{repo_dir}{The package repository directory. Defaults to \code{"./repo"}.}

\item{lib_dir}{Package library output directory. Defaults to \code{"./lib"}.}

\item{strip}{A character vector of directories to strip from each R package.}
}
\description{
Extracts all R packages contained in the repository directory \code{repo_dir} and
writes them to a library directory \code{lib_dir}.
}
\details{
The \code{lib_dir} directory will be created if it does not already exist.

The \code{strip} argument may be used to strip certain directories from packages
installed to the library directory \code{lib_dir}. This can be used to reduce the
total library file size by removing directories that are not strictly
necessary for the R package to run, such as directories containing
documentation and vignettes.
}
