% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_vfs_image}
\alias{make_vfs_image}
\title{Build an Emscripten VFS image containing all the binary R packages in the
given CRAN-like repository.}
\usage{
make_vfs_image(out_dir = "./vfs", ...)
}
\arguments{
\item{out_dir}{The output directory for the result VFS image files.}

\item{...}{Additional arguments passed to \link{make_library}.}
}
\value{
The status code as returned by Emscripten's \code{file_packager} tool.
}
\description{
This creates a single Emscripten VFS image of an R library that contains all
the binary R packages in the given repository. This image can be statically
hosted on a web server and downloaded at runtime by webR as an efficient way
to provide a pre-configured R library without installing each R package
individually.
}
