% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_vfs_repo}
\alias{make_vfs_repo}
\title{Add Emscripten VFS images to a given CRAN-like repo}
\usage{
make_vfs_repo(repo_dir = "./repo")
}
\arguments{
\item{repo_dir}{The CRAN-like repository directory.}
}
\description{
Emscripten VFS images may be used by webR to download and install R packages
faster by mounting images to the VFS, rather than decompressing and
extracting \code{tar} files.
}
