% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm.R
\docType{package}
\name{rwasm-package}
\alias{rwasm}
\alias{rwasm-package}
\title{rwasm: Build R Packages for WebAssembly}
\description{
Utilities to help build R packages compiled for WebAssembly, manage binary WebAssembly R package libraries and repositories, and prepare webR compatible filesystem images for static web hosting of selected data files and R packages.
}
\author{
\strong{Maintainer}: George Stagg \email{george.stagg@posit.co}

Other contributors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co} [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
