% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{write_packages}
\alias{write_packages}
\title{Write the \code{PACKAGES} file for a package repository}
\usage{
write_packages(repo_dir = "./repo")
}
\arguments{
\item{repo_dir}{The package repository directory. Defaults to \code{"./repo"}.}
}
\description{
Write the \code{PACKAGES} file for a package repository
}
