% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_list}
\alias{add_list}
\title{Add one or more packages from a file}
\usage{
add_list(list_file, ...)
}
\arguments{
\item{list_file}{Path to a file containing a list of R package references.}

\item{...}{
  Arguments passed on to \code{\link[=add_pkg]{add_pkg}}
  \describe{
    \item{\code{repo_dir}}{The package repository directory. Defaults to \code{"./repo"}.}
    \item{\code{remotes}}{A character vector of package references to prefer as a remote
source. Defaults to \code{NA}, meaning prefer a built-in list of references to
packages pre-modified for use with webR.}
    \item{\code{dependencies}}{Dependency specification for packages to additionally
add to the repository. Defaults to \code{NA}, meaning add only the required
dependencies. If \code{FALSE}, adds no additional packages. See
\link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies} for details.}
  }}
}
\description{
Downloads and builds the list of \link[pkgdepends:pkg_refs]{R package references}
in the file \code{list_file}, compiling each package for use with WebAssembly and
webR. The resulting WebAssembly binary packages are added to the repository
directory \code{repo_dir}. The repository directory will be created if it does not
already exist.
}
\details{
The R package references should be listed in the file \code{list_file}, one line
per package reference.
}
