% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_pkg}
\alias{add_pkg}
\title{Add R package reference(s) to a package repository}
\usage{
add_pkg(
  packages,
  repo_dir = "./repo",
  remotes = NA,
  dependencies = FALSE,
  compress = TRUE
)
}
\arguments{
\item{packages}{A character vector of one or more package references.}

\item{repo_dir}{The package repository directory. Defaults to \code{"./repo"}.}

\item{remotes}{A character vector of package references to prefer as a remote
source. Defaults to \code{NA}, meaning prefer a built-in list of references to
packages pre-modified for use with webR.}

\item{dependencies}{Dependency specification for packages to additionally
add to the repository. Defaults to \code{FALSE}, meaning no additional packages.
Use \code{NA} to install only hard dependencies whereas \code{TRUE} installs all
optional dependencies as well. See \link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies}
for details.}

\item{compress}{When \code{TRUE}, add and compress Emscripten virtual filesystem
metadata in the resulting R package binary \code{.tgz} files. Otherwise,
\code{\link[=file_packager]{file_packager()}} is used to create uncompressed virtual filesystem images
included in the output binary package repository. Defaults to \code{TRUE}.}
}
\description{
Downloads and builds the \link[pkgdepends:pkg_refs]{R package references} given
by \code{packages}, compiling each package for use with WebAssembly and webR. The
resulting WebAssembly binary packages are added to the repository directory
\code{repo_dir}. The repository directory will be created if it does not already
exist.
}
