% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_pkg}
\alias{add_pkg}
\title{Add packages to a CRAN-like repository}
\usage{
add_pkg(packages, remotes = NULL, repo_dir = "./repo")
}
\arguments{
\item{packages}{A character vector of one or more package references.}

\item{remotes}{A character vector of package references to prefer as a given
package's remote source. If \code{NULL}, use a built-in list of references to
packages pre-modified for use with webR.}

\item{repo_dir}{The CRAN-like repository directory. Will be created if it
does not exist.}
}
\description{
Add packages to a CRAN-like repository
}
