% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{add_repo}
\alias{add_repo}
\title{Add all available packages from a CRAN-like repository to a Wasm repository}
\usage{
add_repo(repos = getOption("repos"), ...)
}
\arguments{
\item{repos}{The base URL(s) of the repositories to search for packages.}

\item{...}{
  Arguments passed on to \code{\link[=add_pkg]{add_pkg}}
  \describe{
    \item{\code{remotes}}{A character vector of package references to prefer as a given
package's remote source. If \code{NULL}, use a built-in list of references to
packages pre-modified for use with webR.}
    \item{\code{repo_dir}}{The Wasm repository directory. Will be created if it does
not exist.}
  }}
}
\description{
Add all available packages from a CRAN-like repository to a Wasm repository
}
