% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build one or more R packages for WebAssembly}
\usage{
build(
  packages,
  out_dir = ".",
  remotes = NULL,
  dependencies = FALSE,
  compress = FALSE
)
}
\arguments{
\item{packages}{A character vector of one or more package references.}

\item{out_dir}{The output directory. Defaults to \code{"."}.}

\item{remotes}{A character vector of package references to prefer as a remote
source. If \code{NA}, use a built-in list of references to packages pre-modified
for use with webR. Defaults to \code{NULL}, meaning no preference over the usual
remote sources.}

\item{dependencies}{Dependency specification for packages to additionally
add to the repository. Defaults to \code{FALSE}, meaning no additional packages.
Use \code{NA} to install only hard dependencies whereas \code{TRUE} installs all
optional dependencies as well. See \link[pkgdepends:as_pkg_dependencies]{pkgdepends::as_pkg_dependencies}
for details.}

\item{compress}{Logical. If \code{TRUE}, a compressed version of the filesystem
data is included in the output. Defaults to \code{FALSE}.}
}
\description{
Downloads and builds the \link[pkgdepends:pkg_refs]{R package references} given
by \code{packages}, compiling each package for use with WebAssembly and webR. The
resulting WebAssembly binary packages are written to \code{out_dir}.
}
