% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_library}
\alias{make_library}
\title{Make an R library directory containing all the binary R packages in the given
CRAN-like repository.}
\usage{
make_library(repo_dir = "./repo", lib_dir = "./lib", strip = NULL)
}
\arguments{
\item{repo_dir}{The CRAN-like repository directory.}

\item{lib_dir}{The library directory. Will be created if it does not exist.}

\item{strip}{A character vector of directories to strip from the packages.}
}
\description{
The \code{strip} argument may be used to strip certain directories from R packages
installed to the library given by \code{lib_dir}. This can be used to reduce the
total library file size by removing directories that are not strictly
necessary for the R package to run, such as directories containing
documentation and vignettes.
}
