% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar.R
\name{make_tar_index}
\alias{make_tar_index}
\title{Create an Emscripten metadata file for a given \code{tar} archive}
\usage{
make_tar_index(file, strip = 0)
}
\arguments{
\item{file}{Filename of the \code{tar} archive to be used as input.}

\item{strip}{Remove the specified number of leading path elements. Pathnames
with fewer elements are skipped. Defaults to \code{0}, meaning none.}
}
\description{
Calculates file offsets and other metadata for content stored in an
(optionally gzip compressed) \code{tar} archive. Together the \code{tar} archive and
resulting metadata file can be mounted as an Emscripten filesystem image,
making the content of the archive available to the WebAssembly R process.
}
\details{
Outputs a metadata file named by the base name of the \code{tar} archive with new
extension \code{".js.metadata"}. Both files should be hosted online so that their
URL can be provided to webR for mounting on the virtual filesystem.
}
