% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_vfs_library}
\alias{make_vfs_library}
\title{Create an Emscripten filesystem image of an R package library}
\usage{
make_vfs_library(
  out_dir = "./vfs",
  out_name = "library.data",
  repo_dir = "./repo",
  compress = FALSE,
  ...
)
}
\arguments{
\item{out_dir}{Directory in which to write the output image files. Defaults
to \code{"./vfs"}.}

\item{out_name}{A character string for the output library image filename.}

\item{repo_dir}{The package repository directory. Defaults to \code{"./repo"}.}

\item{compress}{When \code{TRUE}, add and compress Emscripten virtual filesystem
metadata in the resulting R package binary \code{.tgz} files. Otherwise,
\code{\link[=file_packager]{file_packager()}} is used to create uncompressed virtual filesystem images
included in the output binary package repository. Defaults to \code{TRUE}.}

\item{...}{
  Arguments passed on to \code{\link[=make_library]{make_library}}
  \describe{
    \item{\code{strip}}{A character vector of directories to strip from each R package.}
  }}
}
\description{
Extracts all binary R packages contained in the repository directory
\code{repo_dir} and creates an Emscripten filesystem image containing the
resulting package library.
}
\details{
A single filesystem image is generated using Emscripten's \code{\link[=file_packager]{file_packager()}}
tool and the output \code{.data} and \code{.js.metadata} filesystem image files are
written to the directory \code{out_dir}.

When \code{compress} is \code{TRUE}, an additional file with extension \code{".data.gz"} is
also output containing a compressed version of the filesystem data.

The resulting image can be downloaded by webR and mounted on the Emscripten
virtual filesystem as an efficient way to provide a pre-configured R library,
without installing each R package individually.
}
