% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_vfs_repo}
\alias{make_vfs_repo}
\title{Add Emscripten filesystem images to an R package repository}
\usage{
make_vfs_repo(repo_dir = "./repo", compress = FALSE)
}
\arguments{
\item{repo_dir}{The package repository directory. Defaults to \code{"./repo"}.}

\item{compress}{If \code{TRUE}, use R binary \code{.tgz} files when creating Emscripten
filesystem image metadata. Otherwise, an additional \code{.data} filesystem image
file is created by \code{\link[=file_packager]{file_packager()}} and included in the output repository
binary directory. Defaults to \code{TRUE}.}
}
\description{
Creates an Emscripten filesystem image for each R package that exists in the
package repository directory \code{repo_dir}.
}
\details{
Each filesystem image is generated using Emscripten's \code{\link[=file_packager]{file_packager()}} tool
and the output \code{.data} and \code{.js.metadata} filesystem image files are written
to the repository in the same directory as the package binary \code{.tgz} files.

The resulting filesystem images may then be used by webR to download and
install R packages by mounting the \code{.data} images to the Emscripten virtual
filesystem.

When \code{compress} is \code{TRUE}, an additional file with extension \code{".data.gz"} is
also output containing a compressed version of the filesystem data.
}
