% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_vfs_repo}
\alias{make_vfs_repo}
\title{Add Emscripten filesystem images to an R package repository}
\usage{
make_vfs_repo(repo_dir = "./repo", compress = FALSE)
}
\arguments{
\item{repo_dir}{The package repository directory. Defaults to \code{"./repo"}.}

\item{compress}{Logical. If \code{TRUE}, a compressed version of the filesystem
data is included in the output. Defaults to \code{FALSE}.}
}
\description{
Creates an Emscripten filesystem image for each R package that exists in the
package repository directory \code{repo_dir}.
}
\details{
Each filesystem image is generated using Emscripten's \code{\link[=file_packager]{file_packager()}} tool
and the output \code{.data} and \code{.js.metadata} filesystem image files are written
to the repository in the same directory as the package binary \code{.tar.gz}
files.

The resulting filesystem images may then be used by webR to download and
install R packages faster by mounting the \code{.data} images to the Emscripten
virtual filesystem, rather than decompressing and extracting the equivalent
\code{.tar.gz} files.
}
