% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{rm_pkg}
\alias{rm_pkg}
\title{Remove packages from a CRAN-like repository}
\usage{
rm_pkg(packages, repo_dir = "./repo")
}
\arguments{
\item{packages}{A character vector of one or more package names.}

\item{repo_dir}{The CRAN-like repository directory.}
}
\description{
Remove packages from a CRAN-like repository
}
