% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwasm.R
\docType{package}
\name{rwasm-package}
\alias{rwasm}
\alias{rwasm-package}
\title{rwasm: Build R Packages for WebAssembly}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to help build R packages compiled for WebAssembly (Wasm), manage Wasm binary R package libraries and repositories, and prepare webR compatible filesystem images for static web hosting of data files and R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-wasm.github.io/rwasm/}
  \item \url{https://github.com/r-wasm/rwasm}
  \item Report bugs at \url{https://github.com/r-wasm/rwasm/issues}
}

}
\author{
\strong{Maintainer}: George Stagg \email{george.stagg@posit.co}

Other contributors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co} [contributor]
  \item Jeroen Ooms \email{jeroen@berkeley.edu} [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
